+++ pkg/pantheon/pantheon.go
@@ -5,24 +5,19 @@ import (
 	"math/rand"
 
 	"github.com/ironarachne/world/pkg/language"
+	"github.com/ironarachne/world/pkg/pantheon/deity"
+	"github.com/ironarachne/world/pkg/pantheon/domain"
 )
 
 // Pantheon is a nonhierarchical group of deities
 type Pantheon struct {
-	Deities map[string]Deity
-}
-
-// SimplifiedPantheon is a simplified version of the data for display
-type SimplifiedPantheon struct {
-	Deities []SimplifiedDeity `json:"deities"`
+	Deities []deity.Deity
 }
 
 // Generate creates a random pantheon of deities
 func Generate(minSize int, maxSize int, lang language.Language) (Pantheon, error) {
 	var pantheon Pantheon
 	numberOfDeities := minSize
-
-	pantheon.Deities = make(map[string]Deity)
 
 	if maxSize-minSize > 0 {
 		numberOfDeities = rand.Intn(maxSize-minSize) + minSize
@@ -32,13 +27,16 @@ func Generate(minSize int, maxSize int, lang language.Language) (Pantheon, error
 		return pantheon, nil
 	}
 
+	possibleDomains := domain.All()
+
 	for i := 0; i < numberOfDeities; i++ {
-		deity, err := pantheon.GenerateDeity(lang)
+		d, err := deity.Generate(lang, possibleDomains)
 		if err != nil {
 			err = fmt.Errorf("Could not generate pantheon: %w", err)
 			return Pantheon{}, err
 		}
-		pantheon.Deities[deity.Name] = deity
+		pantheon.Deities = append(pantheon.Deities, d)
+		possibleDomains = domain.Exclude(d.Domains, possibleDomains)
 	}
 
 	if len(pantheon.Deities) > 1 {
@@ -51,17 +49,4 @@ func Generate(minSize int, maxSize int, lang language.Language) (Pantheon, error
 	}
 
 	return pantheon, nil
-}
-
-// Simplify returns a simplified pantheon for display
-func (pantheon Pantheon) Simplify() SimplifiedPantheon {
-	sp := SimplifiedPantheon{}
-	simplifiedDeity := SimplifiedDeity{}
-
-	for _, d := range pantheon.Deities {
-		simplifiedDeity = d.Simplify()
-		sp.Deities = append(sp.Deities, simplifiedDeity)
-	}
-
-	return sp
 }